Fix required xp to next lvl

+++ Scenes/World/Entities/Character/Player/Player.cs
@@ -13,8 +13,8 @@ public partial class Player : Character
 {
 	[Export] [NotNull] public Sprite2D ShieldSprite { get; private set; }
 	
-	public int Xp { get; set; }
-	public int NextLevelXp { get; set; }
+	public long Xp { get; set; }
+	public long NextLevelXp { get; set; }
 	public double RequiredXpLevelFactor { get; set; } = 1.5;
 	public int BasicRequiredXp { get; set; } = 10;
 	public int Level { get; set; } = 1;
